+++ src/test/java/za/co/hpsc/web/controllers/IpscCompetitorControllerTest.java
@@ -94,7 +94,7 @@ class IpscCompetitorControllerTest {
 
     // createCompetitors()
     private static final String VALID_CSV = """
-            FirstName,LastName,MiddleNames,Nickname,DateOfBirth,Gender,HomeClub,SapsaNumber,CompetitorNumber,ClubNumber,IdNumber,CellphoneNumber,EmailAddresses
+            FirstName,LastName,MiddleNames,Nickname,DateOfBirth,Gender,HomeClub,SapsaNumber,CompetitorNumber,ClubNumber,IdNumber,CellphoneNumber,EmailAddresses,PaidUpSapsa,PaidUpClub
             John,Doe,,,,,,,,CLUB001,,,
             """;